Remove the singleton restriction and introduce the concept of "instance relevant package"
Now the package works also when not in use as the sole full-text-search driver on the system. Searchable packages that are "relevant" will be indexed regardless.
Currently, a package is relevtn if it belongs to the subtree of the ollama instance.

A basic UI has been introduced to link knowledge base packages on the index page.